added tests for scraper code
