+++ backend/scraper/code/profile.py
@@ -1,6 +1,6 @@
 import json
 import sys
-from stackOverflowScraper import StackOverflowProfile
+from backend.scraper.code.stackOverflowScraper import StackOverflowProfile
 
 
 class UserProfile:
@@ -9,9 +9,11 @@ class UserProfile:
         "takes the user details dictionary and creates object which is composed of scraper objects"
 
         self._username = user_details["username"]
+        self._links = user_details["links"] if "links" in user_details else {}
 
-        self._stack_profile = StackOverflowProfile(
-            user_details["stack_profile"]) if "stack_profile" in user_details else None
+        if len(self._links) > 0:
+            self._stack_profile = StackOverflowProfile(
+                self._links["stack_profile"]) if "stack_profile" in user_details["links"] else None
 
     def get_next_stack_tag(self):
         return self._stack_profile.getTopTags().__next__() if self._stack_profile is not None else None
@@ -26,19 +28,10 @@ class UserProfile:
         return self._username
 
 
-# todo create more comprehensive tests
-def test_user_profile():
-    with open("/scraper/profile_test.json") as json_file:
-        data = json.load(json_file)
-
-        UserProfile(json.load(json_file))
-
-
 if __name__ == "__main__":
     input_json = json.loads(sys.argv[1])
     profile = UserProfile(input_json)
     output = {
-            'text': profile.get_next_asked_stack_post().getPost()
+        'text': profile.get_next_asked_stack_post().getPost()
     }
     sys.stdout.write(json.dumps(output))
-
