+++ backend/scraper/code/tests/profile_test.json
@@ -1,0 +1,20 @@
+{
+  "test1": {
+    "username": "test user 1",
+    "links": {
+      "stack_profile": "https://stackoverflow.com/users/12870/oli"
+    }
+  },
+  "test2": {
+    "username": "test user 2",
+    "links": {
+      "stack_profile": "https://stackoverflow.com/users/87234/gmannickg"
+    }
+  },
+  "test3": {
+    "username": "test user 3",
+    "links": {
+      "stack_profile": "https://stackoverflow.com/users/6068/pupeno"
+    }
+  }
+}
