--- backend/scraper/profile_test.json
@@ -1,8 +1,0 @@
-{
-  "olo": [
-    {
-      "user_name": "oli1";,
-      "stack_profile": "https://stackoverflow.com/users/12870/oli"
-    }
-  ]
-}
